requirements.txt
pip~=25.1.1
pytz~=2025.2
numpy~=2.3.3
pandas~=2.3.2
python-dateutil~=2.9.0.post0
six~=1.17.0
psycopg2-binary~=2.9.10